+++ app/api/catalogs/[catalogId]/update/route.ts
@@ -2,6 +2,7 @@ import { NxResponse } from "@/lib/nx-response";
 import { getUserIdCookie } from "@/lib/server-helper";
 import { NextRequest } from "next/server";
 import { updateCatalogVideos, updateChannels } from "../../models";
+import { revalidatePath } from "next/cache";
 
 type ContextParams = {
   params: {
@@ -10,21 +11,21 @@ type ContextParams = {
 };
 
 export async function GET(_request: NextRequest, ctx: ContextParams) {
-    const { catalogId } = ctx.params;
-    const data = await updateCatalogVideos(catalogId);
+  const { catalogId } = ctx.params;
+  const data = await updateCatalogVideos(catalogId);
 
-    if (typeof data === "string") {
-      return NxResponse.fail(data, { code: "UNKOWN", details: data }, 404);
-    }
+  if (typeof data === "string") {
+    return NxResponse.fail(data, { code: "UNKOWN", details: data }, 404);
+  }
 
-    return NxResponse.success<{ videos: any; nextUpdate: string }>(
-      "Catalog page updated successfully.",
-      {
-        videos: data.videos,
-        nextUpdate: data.nextUpdate,
-      },
-      200
-    );
+  return NxResponse.success<{ videos: any; nextUpdate: string }>(
+    "Catalog page updated successfully.",
+    {
+      videos: data.videos,
+      nextUpdate: data.nextUpdate,
+    },
+    200
+  );
 }
 
 export async function PATCH(request: NextRequest, ctx: ContextParams) {
@@ -34,7 +35,10 @@ export async function PATCH(request: NextRequest, ctx: ContextParams) {
   if (!catalogId) {
     return NxResponse.fail(
       "Catalog ID is missing from request params.",
-      { code: "BAD_REQUEST", details: "Catalog ID is missing from request params." },
+      {
+        code: "BAD_REQUEST",
+        details: "Catalog ID is missing from request params.",
+      },
       400
     );
   }
@@ -44,7 +48,10 @@ export async function PATCH(request: NextRequest, ctx: ContextParams) {
   await updateChannels(userId, catalogId, catalogPayload);
 
   // Update the catalog
-  await updateCatalogVideos(catalogId);
+  updateCatalogVideos(catalogId);
+
+  // Revalidate the /explore route
+  revalidatePath("/explore");
 
   return NxResponse.success<any>("Channel list update successfully.", {}, 201);
 }